Validate an existing humanoid parent. Fixes 2 reported Engine Bug reports and a logger warning.
Fixes: https://devforum.roblox.com/t/playermove-called-but-player-currently-has-no-humanoid-fix/558942

And potentially:

https://devforum.roblox.com/t/parenting-character-to-nil-then-parenting-to-workspace-makes-player-move-to-000/1452195

This validation will not run the RenderStepped contents, specifically the move function if humanoid parent is nil. Otherwise this results in a warning from the engine and lag the console.

Previously this also caused replication issues where the move gets stuck to ZERO VECTOR for whatever reason, meaning upon re-setting the humanoid parent to the character server side, humanoid Move is called Vector3.new(0,0,0). I've tried to debug this for a week now, where this issue might potentially come from, but have failed to do so.

This validation is worth a shot.